Updated hashbang and permissions for shell scripts (#75)
Git is tracking file permissions, so using chmod +x in shell scripts is redundant.
To make a shell script executable, it's enough to execute chmod +x and commit the file. On Windows, it's possible to use the next command:
git update-index --chmod=+x <path_to_shell>

The source command is not compatible with sh specification, thus hash-bang points to bash. All scripts touched for consistency.